Enable user code to override struct timeout_cb
I've found that frequently a void function and a pointer are either
more or less than we want to attach to each timeout object.  This
patch makes it so that the user can define TIMEOUT_CB_OVERRIDE to
indicate that they have included their own definition of struct
timeout_cb.