Make the 'timeouts' field in struct timeout optional.
This field eats a pointer for every struct timeout, and it only
exists in order to enable a set of APIs that omit the 'struct
timeouts' object. For some applications, the convenience of these
APIs is not worth the memory overhead. This patch makes it so
that applications can disable those APIs, and save some memory, by
defining TIMEOUT_DISABLE_RELATIVE_ACCESS at compile time.