Small optimization: remove needless code from internal sched() calls
When we call timeout_sched() internally, sometimes we can be sure
that the timeout is in the future.  Always, we can be sure that the
timeout has its expires field and its timeouts field set correctly,
and that it is not currently pending.

(Do not merge this patch without benchmarking to see if it makes a
measurable difference.)